Introduce RPC connection header
* Update the RPC header protobufs to more closely match Hadoop
* Switch callIds to signed 32-bit ints to match Hadoop
* Add methods to serialize & validate connection headers
* Add unit tests for connection header serialization/validation
* Update docs to describe connection header
* Add a comment in protoc-gen-krpc.cc showing usage

Reviewed-on: http://gerrit.ent.cloudera.com:8080/293
Tested-by: jenkins

+++ src/rpc/server_call.h
@@ -48,7 +48,7 @@ class InboundCall {
     return header_.methodname();
   }
 
-  const uint64_t call_id() const {
+  const int32_t call_id() const {
     return header_.callid();
   }
